refactor(strands): split session manager into focused modules
session_manager.py keeps only the SessionManager protocol implementation
(lifecycle, hooks, write-behind buffer). Pure message mapping moves to
_messages.py; Neo4jRetrievalConfig + search/formatting move to
_retrieval.py (decoupled from session state — takes long_term + config).
Mirrors the mcp package's _tools.py/_resources.py convention.
